+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.1
+
+* Add `TiledMap.getObjectTileCoordinates`
+* Add `TiledMap.getObjectByName`
+
 ## 0.1.0
 
 * :rocket: Initial version